requirements.txt
Flask==3.0.0
python-docx==1.1.2